don't show defiance damage on facts with special text
We do this because most like facts with special text have a special meaning and dont just change the name; for example a fact might remove imob and override the text of the buff fact with 'Removes Condition'. In those cases we dont want to process the defiance damage. Might be unstable, just testing it for now.

Also added a listing in the individual tests for it.