- added item salvage types - we can now use an actual api.
The new data acquisition order for tests is local api > global objects > > synthesize > error.

I'll probably remove the global objects at some point, but for now they are a convenient way to work on the renderer without an api.
MAybe ill also go the other route and add back an exporter to the api.