added filtering for dev icons to trait sub-skills and completely hide sub skill icons of relics
closes #55

I really want to keep the icons for trait sub-skills as most of them do work and help to add information to the blocks, but sometimes it just doesn't work properly.
Same issue with Relics, however there the icons aren't set for the largest part of the skills we can obtain, so no point in showing them or even trying to filter them.